+++ test/conftest.py
@@ -4,8 +4,7 @@ import pytest
 from os.path import join
 from asv import config
 from asv import repo
-from .test_workflow import generate_basic_conf
-
+from .test_workflow import basic_conf, generate_basic_conf
 
 try:
     import hglib
@@ -114,8 +113,19 @@ def two_branch_repo_case(request, tmpdir):
 
 
 @pytest.fixture
+def basic_conf(tmpdir, dummy_packages):
+    return generate_basic_conf(tmpdir)
+
+
+@pytest.fixture
+def basic_conf_with_subdir(tmpdir, dummy_packages):
+    return generate_basic_conf(tmpdir, 'some_subdir')
+
+
+@pytest.fixture
 def existing_env_conf(tmpdir):
     tmpdir, local, conf, machine_file = generate_basic_conf(tmpdir)
     conf.environment_type = "existing"
     conf.pythons = ["same"]
     return tmpdir, local, conf, machine_file
+
