+++ test/test_run.py
@@ -18,17 +18,6 @@ from asv.commands.run import Run
 
 from . import tools
 from .tools import dummy_packages, WIN
-from .test_workflow import basic_conf, generate_basic_conf
-
-
-@pytest.fixture
-def basic_conf(tmpdir, dummy_packages):
-    return generate_basic_conf(tmpdir)
-
-
-@pytest.fixture
-def basic_conf_with_subdir(tmpdir, dummy_packages):
-    return generate_basic_conf(tmpdir, 'some_subdir')
 
 
 def test_set_commit_hash(capsys, existing_env_conf):
